+++ example/get_android_dependencies.ps1
@@ -1,8 +1,12 @@
-﻿$currentDir = (Get-Item -Path ".\" -Verbose).FullName
+﻿$AneVersion = "1.6.0"
+$PlayerServicesVersion = "15.0.1"
+$SupportV4Version = "27.1.0"
+
+$currentDir = (Get-Item -Path ".\" -Verbose).FullName
 [Net.ServicePointManager]::SecurityProtocol = [Net.SecurityProtocolType]::Tls12
-Invoke-WebRequest -Uri https://github.com/tuarua/AdMob-ANE/releases/download/1.6.0/AdMobANE.ane?raw=true -OutFile "$currentDir\..\native_extension\ane\AdMobANE.ane"
+Invoke-WebRequest -Uri https://github.com/tuarua/AdMob-ANE/releases/download/$AneVersion/AdMobANE.ane?raw=true -OutFile "$currentDir\..\native_extension\ane\AdMobANE.ane"
 Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.tuarua.frekotlin.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/kotlin/com.tuarua.frekotlin.ane?raw=true
 Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.code.gson.gson-2.8.1.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/misc/com.google.code.gson.gson-2.8.1.ane?raw=true
-Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.android.support.support-v4-27.1.0.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/support/com.android.support.support-v4-27.1.0.ane?raw=true
-Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-base-12.0.1.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-base-12.0.1.ane?raw=true
-Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-ads-lite-12.0.1.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-ads-lite-12.0.1.ane?raw=true
+Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.android.support.support-v4-$SupportV4Version.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/support/com.android.support.support-v4-$SupportV4Version.ane?raw=true
+Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-base-$PlayerServicesVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-base-$PlayerServicesVersion.ane?raw=true
+Invoke-WebRequest -OutFile "$currentDir\android_dependencies\com.google.android.gms.play-services-ads-lite-$PlayerServicesVersion.ane" -Uri https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/play-services/com.google.android.gms.play-services-ads-lite-$PlayerServicesVersion.ane?raw=true