+++ example/get_ios_dependencies.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+AneVersion="1.6.0"
+PlayerServicesVersion="15.0.1"
+SupportV4Version="27.1.0"
+
 rm -r ios_dependencies/device
 rm -r ios_dependencies/simulator
 
@@ -7,8 +11,8 @@ wget https://github.com/tuarua/Swift-IOS-ANE/releases/download/2.4.0/ios_depende
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget https://github.com/tuarua/AdMob-ANE/releases/download/1.6.0/ios_dependencies.zip
+wget https://github.com/tuarua/AdMob-ANE/releases/download/$AneVersion/ios_dependencies.zip
 unzip -u -o ios_dependencies.zip
 rm ios_dependencies.zip
 
-wget -O ../native_extension/ane/AdMobANE.ane https://github.com/tuarua/AdMob-ANE/releases/download/1.6.0/AdMobANE.ane?raw=true
+wget -O ../native_extension/ane/AdMobANE.ane https://github.com/tuarua/AdMob-ANE/releases/download/$AneVersion/AdMobANE.ane?raw=true